requirements.txt
streamlit
langchain
fastapi[all]
openai
azure-identity
# azure-search-documents==11.4.0b12
./whl/azure_search_documents-11.4.0b12-py3-none-any.whl
azure-storage-blob
python-dotenv
azure-cosmos==4.5.2b3
requests
tiktoken
pymupdf
azure-monitor-opentelemetry
requests
unstructured[all-docs]